refactor: pull out expandable accordion component
